agenteinsights/requirements.txt
pandas
openpyxl
scikit-learn
matplotlib
seaborn
python-docx
python-dotenv
openai